Reverting attempt to move player code into it's own class.
